Moved DataSources to their own folder.
